+++ src/Git_Remote_Updater/Settings.php
@@ -88,7 +88,7 @@ class Settings {
 		esc_html_e( 'Git Remote Updater', 'git-remote-updater' );
 		echo '</h2>';
 
-		if ( 'git-remote-updater' == $page ) {
+		if ( 'git-remote-updater' === $page ) {
 			( new Actions() )->display( $action );
 		}
 		if ( 'git-remote-updater-settings' === $page ) {
@@ -119,14 +119,15 @@ class Settings {
 				'api_key' => $_POST['git_remote_updater_key'],
 			];
 			$new_options = $this->sanitize( $new_options, $options );
+			$empty_add   = empty( $_POST['git_remote_updater_site'] );
 
 			foreach ( $options as $option ) {
 				$duplicate = in_array( $new_options[0]['ID'], $option, true );
 				if ( $duplicate ) {
-					continue;
+					break;
 				}
 			}
-			if ( ! $duplicate ) {
+			if ( ! $duplicate && ! $empty_add ) {
 				$options = array_merge( $options, $new_options );
 				update_site_option( 'git_remote_updater', $options );
 			}